_AutoTriageScripts/requirements.txt
requests>=2.31.0
openai>=1.12.0  # For OpenAI-compatible API interfaces
python-dotenv>=1.0.0  # For environment variable management
pyyaml>=6.0.1  # For YAML file parsing